spring-boot-book-jdbc/src/main/java/com/springboot/book/jdbc/book/beans/Book.java
package com.springboot.book.jdbc.book.beans;

import org.springframework.stereotype.Component;

@Component
public class Book {
	private int bookId;
	private String title;
	private String author;
	private double price;
	private String publications;
	private int year;
	public int getBookId() {
		return bookId;
	}
	public void setBookId(int bookId) {
		this.bookId = bookId;
	}
	public String getTitle() {
		return title;
	}
	public void setTitle(String title) {
		this.title = title;
	}
	public String getAuthor() {
		return author;
	}
	public void setAuthor(String author) {
		this.author = author;
	}
	public double getPrice() {
		return price;
	}
	public void setPrice(double price) {
		this.price = price;
	}
	public String getPublications() {
		return publications;
	}
	public void setPublications(String publications) {
		this.publications = publications;
	}
	public int getYear() {
		return year;
	}
	public void setYear(int year) {
		this.year = year;
	}
}
